Changed title_id to bigint.

+++ CreateTables/8_create_title.sql
@@ -1,5 +1,5 @@
 CREATE TABLE Title(
-    title_id INT NOT NULL AUTO_INCREMENT,
+    title_id BIGINT NOT NULL AUTO_INCREMENT,
   #  recipe_id INT NOT NULL,
     title VARCHAR(100) NOT NULL,
     PRIMARY KEY (title_id)
